agent/requirements.txt
web3==7.6.0
eth-account==0.13.4
python-dotenv==1.0.1
aiohttp==3.10.10
ens==0.1.1
cryptography==43.0.3
rich==13.9.4
